moved UI screens to component & wrote logic in screens to ease testing
