+++ package.json
@@ -17,7 +17,7 @@
     "@react-navigation/native-stack": "^6.9.1",
     "@testing-library/react-native": "^11.4.0",
     "react": "18.1.0",
-    "react-native": "0.70.5",
+    "react-native": "0.70.6",
     "react-native-qrcode-svg": "^6.1.2",
     "react-native-safe-area-context": "^4.4.1",
     "react-native-screens": "^3.18.2",
