+++ src/components/index.ts
@@ -1,4 +1,5 @@
 /* istanbul ignore file */
 import {Login} from './Login';
+import {Profile} from './Profile';
 
-export {Login};
+export {Login, Profile};
