+++ src/screens/ProfileScreen/index.tsx
@@ -1,15 +1,11 @@
 import React, {ReactElement} from 'react';
-
+import {Platform} from 'react-native';
 import {useNavigation} from '@react-navigation/native';
-
-import {TouchableOpacity, SafeAreaView, Text, Platform} from 'react-native';
-
 import {RootStackNavigation} from '../../navigation';
 import auth from '@react-native-firebase/auth';
 import {StackActions} from '@react-navigation/native';
 import analytics from '@react-native-firebase/analytics';
-import QRCode from 'react-native-qrcode-svg';
-import styles from './index.styles';
+import {Profile} from '../../components';
 
 const ProfileScreen = (): ReactElement => {
   const navigation = useNavigation<RootStackNavigation>();
@@ -24,15 +20,10 @@ const ProfileScreen = (): ReactElement => {
   };
 
   return (
-    <SafeAreaView style={styles.container}>
-      <Text>Welcome {auth().currentUser?.displayName || 'User'}</Text>
-
-      <QRCode value="Thanks for logging in!" />
-
-      <TouchableOpacity style={styles.logoutBtnContainer} onPress={onLogout}>
-        <Text style={styles.logoutText}>Logout</Text>
-      </TouchableOpacity>
-    </SafeAreaView>
+    <Profile
+      heading={'Welcome' + auth().currentUser?.email}
+      onLogout={onLogout}
+    />
   );
 };
