--- src/screens/SplashScreen/index.styles.ts
@@ -1,11 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const styles = StyleSheet.create({
-  container: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    flex: 1,
-  },
-});
-
-export default styles;
